use cstdint types

+++ SmoothStepper.h
@@ -50,7 +50,7 @@ class SmoothStepper {
     void setInvertDirection(bool invert);
 
     // set the number of steps per revolution
-    void setStepsPerRev(unsigned long stepsPerRev);
+    void setStepsPerRev(uint32_t stepsPerRev);
 
     // set the acceleration (steps / s^2)
     void setAcceleration(float acceleration);
@@ -59,7 +59,7 @@ class SmoothStepper {
     void setMaxSpeed(float maxSpeed);
 
     // set the duration of step pulses (µs)
-    void setPulseWidth(unsigned int pulseWidth);
+    void setPulseWidth(uint16_t pulseWidth);
 
     // enables the stepper motor  driver (via _pinEnable)
     void enableDriver();
@@ -68,12 +68,12 @@ class SmoothStepper {
     void disableDriver();
 
     // move by n steps
-    void moveSteps(long nSteps);
+    void moveSteps(int32_t nSteps);
 
     // move by n degrees
     void moveDegrees(float degrees);
 
-    
+
   private:
     void step();                      // step function
     uint8_t _pinStep;                 // pin number: step
@@ -82,8 +82,8 @@ class SmoothStepper {
     float _a;                         // acceleration (steps / s^2)
     float _vMax;                      // maximum speed (steps / s)
     float _c0;                        // duration of first interval (µs)
-    unsigned int _pulseWidth = 1;     // duration of step pulses (µs)
-    unsigned long _stepsPerRev = 200; // steps per revolution
+    uint16_t _pulseWidth = 1;         // duration of step pulses (µs)
+    uint32_t _stepsPerRev = 200;      // steps per revolution
     bool _invertEnable = false;       // invert the enable pin?
     bool _invertDirection = false;    // invert the direction pin?
 };
